ACCEL-554: Allows delete assigned cohorts only for staff role.

+++ lms/djangoapps/instructor/views/api.py
@@ -944,7 +944,8 @@ def modify_access(request, course_id):
         allow_access(course, user, rolename)
     elif action == 'revoke':
         revoke_access(course, user, rolename)
-        CohortAssigment.objects.delete_assigned_cohorts(user)
+        if rolename == 'staff':
+            CohortAssigment.objects.delete_assigned_cohorts(user)
     else:
         return HttpResponseBadRequest(strip_tags(
             "unrecognized action '{}'".format(action)
